+++ app/model/user.model.js
@@ -6,7 +6,7 @@ module.exports = (sequelize, Sequelize) => {
 	  Email_address: {type:Sequelize.STRING},
 	  Gst_no: {type:Sequelize.STRING},
 	
-	  phone_no: {type:Sequelize.INTEGER(15)},
+	  phone_no: {type:Sequelize.BIGINT},
 	    Registration_certificate: {type:Sequelize.STRING},  
 		GST_Certificate: {type:Sequelize.STRING},  
 		Pan_card: {type:Sequelize.STRING},
